fix visibleXRange prop

+++ ios/ReactNativeCharts/RNBarLineChartViewBase.swift
@@ -181,8 +181,8 @@ class RNBarLineChartViewBase: RNYAxisChartViewBase {
     func setVisibleXRange(_ config: NSDictionary) {
         let json = BridgeUtils.toJson(config)
 
-        let minXRange = json["minXRange"].double != nil ? CGFloat(json["minXRange"].doubleValue) : 1
-        let maxXRange = json["maxXRange"].double != nil ? CGFloat(json["maxXRange"].doubleValue) : 1
+        let minXRange = json["minXRange"].double != nil ? json["minXRange"].doubleValue : 1
+        let maxXRange = json["maxXRange"].double != nil ? json["maxXRange"].doubleValue : 1
 
         barLineChart.setVisibleXRange(minXRange: minXRange, maxXRange: maxXRange)
     }
